rename interfaces to rpc_api
